Export mixin helper functions separately
